Replace ChromeTune app struct name with App
The cli code I copied from the Crux repo was complaining that
shared::App didn't exist.

This was due to a small discrepency between the example I copied from [the Crux
book](https://redbadger.github.io/crux/getting_started/core.html#the-app),
and the Counter example in the crux repo.

Since the majority of examples in the repo uses App identifier, and
since App is a bit more generic, and less prone to make the same mistake
in the future if I copy it again, sticking with App.