Fix an off-by-one error
I realized the mistake when I tested it with A2.wav file. Despite the
frequency detected by the `pitch` crate was correct (110 Hz), we were
detecting this as the next string (D3). This was because when we break
from the loop, we already pass the index that was smaller than the
detected Hz value, hence the idx and idx+1 were always going to be
greater than the detected value. The correct range is idx-1 and idx.

Also, to prevent index out of bounds, added a clamp function (copied
from num crate - it was simple enough to include, perhaps it is better
to just use num crate).