Add an alternative crate (pitch-detector) for pitch detection
Since they use different methods (`pitch` uses Bitstream Autocorrelation,
and `pitch-detector` is configurable but currently we set to HannedFft),
having two at this stage is helping me to compare the results.

So far, `pitch` crate seems to be better for the test cases I used, but
since I only used two recordings, it is too early to tell, so there's
value having both.