Add a simple test case in the cli package
This mainly tests the pitch crate, but that doesn't seem to have much
tests either, so having this here is useful. It also gives us confidence
to swap the crate/algorithm later.

I added the test to cli package instead of the shared one, because this
one was already depending on the wavers crate.